requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
attrs==25.3.0
cffi==1.17.1
cryptography==45.0.3
frozenlist==1.6.2
gnureadline==8.2.13
idna==3.10
multidict==6.4.4
pillow==11.2.1
propcache==0.3.1
pycparser==2.22
yarl==1.20.0
